Правки в Отрисуй меня полностью. Часть 1
