+++ index.html
@@ -231,7 +231,7 @@
   </template>
   <script src="js/util.js" type="module"></script>
   <script src="js/data.js" type="module"></script>
-  <script src="js/form.js" type="module"></script>
+  <script src="js/popup.js" type="module"></script>
   <script src="js/main.js"></script>
 </body>
 </html>
