+++ js/util.js
@@ -4,6 +4,4 @@ const getRandomNumber = function(min, max) {
   return Math.floor(Math.random() * (max - min + 1)) + min; //Максимум и минимум включаются
 };
 
-export {
-  getRandomNumber
-};
+export {getRandomNumber};
